requirements.txt
fastmcp>=2.13.1
mcp-proxy-for-aws>=1.1.0
boto3>=1.42.0
botocore>=1.42.0
